@@ -1,12 +1,13 @@
 import React, { PropTypes } from 'react';
 import { Link } from 'react-router';
-const InlineSVG = require('react-inlinesvg');
+import classNames from 'classnames';
+import InlineSVG from 'react-inlinesvg';
+
 const playUrl = require('../../../images/play.svg');
 const logoUrl = require('../../../images/p5js-logo.svg');
 const stopUrl = require('../../../images/stop.svg');
 const preferencesUrl = require('../../../images/preferences.svg');
 const editProjectNameUrl = require('../../../images/pencil.svg');
-import classNames from 'classnames';
 
 class Toolbar extends React.Component {
   constructor(props) {
@@ -38,19 +39,19 @@ class Toolbar extends React.Component {
   }
 
   render() {
-    let playButtonClass = classNames({
+    const playButtonClass = classNames({
       'toolbar__play-button': true,
       'toolbar__play-button--selected': this.props.isPlaying
     });
-    let stopButtonClass = classNames({
+    const stopButtonClass = classNames({
       'toolbar__stop-button': true,
       'toolbar__stop-button--selected': !this.props.isPlaying
     });
-    let preferencesButtonClass = classNames({
+    const preferencesButtonClass = classNames({
       'toolbar__preferences-button': true,
       'toolbar__preferences-button--selected': this.props.preferencesIsVisible
     });
-    let nameContainerClass = classNames({
+    const nameContainerClass = classNames({
       'toolbar__project-name-container': true,
       'toolbar__project-name-container--editing': this.props.project.isEditingName
     });
@@ -110,7 +111,7 @@ class Toolbar extends React.Component {
                 e.preventDefault();
                 this.originalProjectName = this.props.project.name;
                 this.props.showEditProjectName();
-                setTimeout(() => this.refs.projectNameInput.focus(), 0);
+                setTimeout(() => this.projectNameInput.focus(), 0);
               }
             }}
           >
@@ -122,7 +123,7 @@ class Toolbar extends React.Component {
             className="toolbar__project-name-input"
             value={this.props.project.name}
             onChange={this.handleProjectNameChange}
-            ref="projectNameInput"
+            ref={(element) => { this.projectNameInput = element; }}
             onBlur={() => {
               this.validateProjectName();
               this.props.hideEditProjectName();
@@ -157,7 +158,6 @@ class Toolbar extends React.Component {
 Toolbar.propTypes = {
   isPlaying: PropTypes.bool.isRequired,
   preferencesIsVisible: PropTypes.bool.isRequired,
-  startSketch: PropTypes.func.isRequired,
   stopSketch: PropTypes.func.isRequired,
   startTextOutput: PropTypes.func.isRequired,
   stopTextOutput: PropTypes.func.isRequired,
@@ -180,6 +180,11 @@ Toolbar.propTypes = {
   saveProject: PropTypes.func.isRequired,
   currentUser: PropTypes.string,
   clearConsole: PropTypes.func.isRequired
+};
+
+Toolbar.defaultProps = {
+  owner: undefined,
+  currentUser: undefined
 };
 
 export default Toolbar;
