@@ -1,10 +1,11 @@
 import React, { PropTypes } from 'react';
 import InlineSVG from 'react-inlinesvg';
+import { browserHistory } from 'react-router';
+
 const exitUrl = require('../../../images/exit.svg');
 const squareLogoUrl = require('../../../images/p5js-square-logo.svg');
 const playUrl = require('../../../images/play.svg');
 const asteriskUrl = require('../../../images/p5-asterisk.svg');
-import { browserHistory } from 'react-router';
 
 class About extends React.Component {
   constructor(props) {
@@ -13,7 +14,7 @@ class About extends React.Component {
   }
 
   componentDidMount() {
-    this.refs.about.focus();
+    this.aboutSection.focus();
   }
 
   closeAboutModal() {
@@ -22,7 +23,7 @@ class About extends React.Component {
 
   render() {
     return (
-      <section className="about" ref="about" tabIndex="0">
+      <section className="about" ref={(element) => { this.aboutSection = element; }} tabIndex="0">
         <header className="about__header">
           <h2 className="about__header-title">Welcome</h2>
           <button className="about__exit-button" onClick={this.closeAboutModal}>
@@ -36,6 +37,7 @@ class About extends React.Component {
               <a
                 href="http://hello.p5js.org/"
                 target="_blank"
+                rel="noopener noreferrer"
               >
                 <InlineSVG className="about__play-video-button" src={playUrl} alt="Play Hello Video" />
               Play hello! video</a>
@@ -47,6 +49,7 @@ class About extends React.Component {
               <a
                 href="https://p5js.org/examples/"
                 target="_blank"
+                rel="noopener noreferrer"
               >
                 <InlineSVG className="about__content-column-asterisk" src={asteriskUrl} alt="p5 asterisk" />
               Examples</a>
@@ -55,6 +58,7 @@ class About extends React.Component {
               <a
                 href="https://p5js.org/tutorials/"
                 target="_blank"
+                rel="noopener noreferrer"
               >
                 <InlineSVG className="about__content-column-asterisk" src={asteriskUrl} alt="p5 asterisk" />
               Tutorials</a>
@@ -66,6 +70,7 @@ class About extends React.Component {
               <a
                 href="https://p5js.org/libraries/"
                 target="_blank"
+                rel="noopener noreferrer"
               >
                 <InlineSVG className="about__content-column-asterisk" src={asteriskUrl} alt="p5 asterisk" />
               Libraries</a>
@@ -74,6 +79,7 @@ class About extends React.Component {
               <a
                 href="https://p5js.org/reference/"
                 target="_blank"
+                rel="noopener noreferrer"
               >
                 <InlineSVG className="about__content-column-asterisk" src={asteriskUrl} alt="p5 asterisk" />
               Reference</a>
@@ -82,6 +88,7 @@ class About extends React.Component {
               <a
                 href="https://forum.processing.org/two/"
                 target="_blank"
+                rel="noopener noreferrer"
               >
                 <InlineSVG className="about__content-column-asterisk" src={asteriskUrl} alt="p5 asterisk" />
               Forum</a>
@@ -93,18 +100,21 @@ class About extends React.Component {
             <a
               href="https://github.com/processing/p5.js-web-editor"
               target="_blank"
+              rel="noopener noreferrer"
             >Contribute</a>
           </p>
           <p className="about__footer-list">
             <a
               href="https://github.com/processing/p5.js-web-editor/issues/new"
               target="_blank"
+              rel="noopener noreferrer"
             >Report a bug</a>
           </p>
           <p className="about__footer-list">
             <a
               href="https://twitter.com/p5xjs?lang=en"
               target="_blank"
+              rel="noopener noreferrer"
             >Twitter</a>
           </p>
           <button className="about__ok-button" onClick={this.closeAboutModal}>OK!</button>
