@@ -1,12 +1,13 @@
 import React, { PropTypes } from 'react';
 import InlineSVG from 'react-inlinesvg';
+import classNames from 'classnames';
+
 const upArrowUrl = require('../../../images/up-arrow.svg');
 const downArrowUrl = require('../../../images/down-arrow.svg');
-import classNames from 'classnames';
 
 class Console extends React.Component {
   componentDidUpdate() {
-    this.refs.console_messages.scrollTop = this.refs.console_messages.scrollHeight;
+    this.consoleMessages.scrollTop = this.consoleMessages.scrollHeight;
   }
 
   render() {
@@ -16,13 +17,13 @@ class Console extends React.Component {
     });
 
     return (
-      <div ref="console" className={consoleClass} role="main" tabIndex="0" title="console">
+      <div className={consoleClass} role="main" tabIndex="0" title="console">
         <div className="preview-console__header">
           <h2 className="preview-console__header-title">Console</h2>
           <div className="preview-console__header-buttons">
-            <a className="preview-console__clear" onClick={this.props.clearConsole} aria-label="clear console">
+            <button className="preview-console__clear" onClick={this.props.clearConsole} aria-label="clear console">
               Clear
-            </a>
+            </button>
             <button className="preview-console__collapse" onClick={this.props.collapseConsole} aria-label="collapse console">
               <InlineSVG src={downArrowUrl} />
             </button>
@@ -31,20 +32,20 @@ class Console extends React.Component {
             </button>
           </div>
         </div>
-        <div ref="console_messages" className="preview-console__messages">
-          {this.props.consoleEvents.map((consoleEvent, index) => {
+        <div ref={(element) => { this.consoleMessages = element; }} className="preview-console__messages">
+          {this.props.consoleEvents.map((consoleEvent) => {
             const args = consoleEvent.arguments;
             const method = consoleEvent.method;
             if (Object.keys(args).length === 0) {
               return (
-                <div key={index} className="preview-console__undefined">
-                  <span key={`${index}-0`}>{'undefined'}</span>
+                <div key={consoleEvent.id} className="preview-console__undefined">
+                  <span key={`${consoleEvent.id}-0`}>{'undefined'}</span>
                 </div>
               );
             }
             return (
-              <div key={index} className={`preview-console__${method}`}>
-                {Object.keys(args).map((key) => <span key={`${index}-${key}`}>{args[key]}</span>)}
+              <div key={consoleEvent.id} className={`preview-console__${method}`}>
+                {Object.keys(args).map(key => <span key={`${consoleEvent.id}-${key}`}>{args[key]}</span>)}
               </div>
             );
           })}
@@ -56,12 +57,18 @@ class Console extends React.Component {
 }
 
 Console.propTypes = {
-  consoleEvents: PropTypes.array,
-  isPlaying: PropTypes.bool.isRequired,
+  consoleEvents: PropTypes.arrayOf(PropTypes.shape({
+    method: PropTypes.string.isRequired,
+    args: PropTypes.arrayOf(PropTypes.string)
+  })),
   isExpanded: PropTypes.bool.isRequired,
   collapseConsole: PropTypes.func.isRequired,
   expandConsole: PropTypes.func.isRequired,
   clearConsole: PropTypes.func.isRequired
+};
+
+Console.defaultProps = {
+  consoleEvents: []
 };
 
 export default Console;
