@@ -16,16 +16,18 @@ export function createFile(req, res) {
     }, (err, updatedProject) => {
       if (err) {
         console.log(err);
-        return res.json({ success: false });
+        res.json({ success: false });
+        return;
       }
       const newFile = updatedProject.files[updatedProject.files.length - 1];
       updatedProject.files.id(req.body.parentId).children.push(newFile.id);
-      updatedProject.save(innerErr => {
+      updatedProject.save((innerErr) => {
         if (innerErr) {
           console.log(innerErr);
-          return res.json({ success: false });
+          res.json({ success: false });
+          return;
         }
-        return res.json(updatedProject.files[updatedProject.files.length - 1]);
+        res.json(updatedProject.files[updatedProject.files.length - 1]);
       });
     });
 }
@@ -38,13 +40,13 @@ function getAllDescendantIds(files, nodeId) {
 }
 
 function deleteMany(files, ids) {
-  ids.forEach(id => {
+  ids.forEach((id) => {
     files.id(id).remove();
   });
 }
 
 function deleteChild(files, parentId, id) {
-  files = files.map((file) => {
+  return files.map((file) => {
     if (file.id === parentId) {
       file.children = file.children.filter(child => child !== id);
       return file;
@@ -57,11 +59,11 @@ export function deleteFile(req, res) {
   Project.findById(req.params.project_id, (err, project) => {
     const idsToDelete = getAllDescendantIds(project.files, req.params.file_id);
     deleteMany(project.files, [req.params.file_id, ...idsToDelete]);
-    deleteChild(project.files, req.query.parentId, req.params.file_id);
+    project.files = deleteChild(project.files, req.query.parentId, req.params.file_id);
     // project.files.id(req.params.file_id).remove();
     // const childrenArray = project.files.id(req.query.parentId).children;
     // project.files.id(req.query.parentId).children = childrenArray.filter(id => id !== req.params.file_id);
-    project.save(innerErr => {
+    project.save((innerErr) => {
       res.json(project.files);
     });
   });
@@ -70,12 +72,14 @@ export function deleteFile(req, res) {
 export function getFileContent(req, res) {
   Project.findById(req.params.project_id, (err, project) => {
     if (err) {
-      return res.status(404).send({ success: false, message: 'Project with that id does not exist.' });
+      res.status(404).send({ success: false, message: 'Project with that id does not exist.' });
+      return;
     }
     const filePath = req.params[0];
     const resolvedFile = resolvePathToFile(filePath, project.files);
     if (!resolvedFile) {
-      return res.status(404).send({ success: false, message: 'File with that name and path does not exist.' });
+      res.status(404).send({ success: false, message: 'File with that name and path does not exist.' });
+      return;
     }
     res.send(resolvedFile.content);
   });
