@@ -3,10 +3,9 @@ import Q from 'q';
 import mongoose from 'mongoose';
 import objectID from 'bson-objectid';
 import shortid from 'shortid';
+import eachSeries from 'async/eachSeries';
 import User from './models/user';
 import Project from './models/project';
-import async from 'async';
-import eachSeries from 'async/eachSeries';
 
 const defaultHTML =
 `<!DOCTYPE html>
@@ -30,8 +29,8 @@ const defaultCSS =
 }
 `;
 
-const client_id = process.env.GITHUB_ID;
-const client_secret = process.env.GITHUB_SECRET;
+const clientId = process.env.GITHUB_ID;
+const clientSecret = process.env.GITHUB_SECRET;
 
 const headers = { 'User-Agent': 'p5js-web-editor/0.0.1' };
 
@@ -41,144 +40,106 @@ mongoose.connection.on('error', () => {
   process.exit(1);
 });
 
-getp5User();
-
-function getp5User() {
-  User.findOne({ username: 'p5' }, (err, user) => {
-    if (err) throw err;
-
-    if (!user) {
-      user = new User({
-        username: 'p5',
-        email: 'p5-examples@gmail.com',
-        password: 'test'
-      });
-      user.save(err => {
-        if (err) throw err;
-        console.log('Created a user p5' + user);
-      });
-    }
-
-    Project.find({ user: user._id }, (err, projects) => {
-      // if there are already some sketches, delete them
-      console.log('Deleting old projects...');
-      projects.forEach(project => {
-        Project.remove({ _id: project._id }, err => {
-          if (err) throw err;
-        });
-      });
-    });
-
-    return getCategories()
-      .then(getSketchesInCategories)
-      .then(getSketchContent)
-      .then(createProjectsInP5user);
-  });
-}
-
 function getCategories() {
-  let categories = [];
+  const categories = [];
   const options = {
-    url: 'https://api.github.com/repos/processing/p5.js-website/contents/dist/assets/examples/en?client_id=' +
-    client_id + '&client_secret=' + client_secret,
+    url: `https://api.github.com/repos/processing/p5.js-website/contents/dist/assets/examples/en?client_id=${
+    clientId}&client_secret=${clientSecret}`,
     method: 'GET',
     headers
   };
-  return rp(options).then(res => {
+  return rp(options).then((res) => {
     const json = JSON.parse(res);
 
-    json.forEach(metadata => {
+    json.forEach((metadata) => {
       let category = '';
-      for (let j = 1; j < metadata.name.split('_').length; j++) {
-        category += metadata.name.split('_')[j] + ' ';
+      for (let j = 1; j < metadata.name.split('_').length; j += 1) {
+        category += `${metadata.name.split('_')[j]} `;
       }
       categories.push({ url: metadata.url, name: category });
     });
 
     return categories;
-  }).catch(err => {
+  }).catch((err) => {
     throw err;
   });
 }
 
 function getSketchesInCategories(categories) {
-  return Q.all(categories.map(category => {
+  return Q.all(categories.map((category) => {
     const options = {
-      url: category.url.replace('?ref=master', '') + '?client_id=' + client_id + '&client_secret=' + client_secret,
+      url: `${category.url.replace('?ref=master', '')}?client_id=${clientId}&client_secret=${clientSecret}`,
       method: 'GET',
       headers
     };
 
-    return rp(options).then(res => {
-      let projectsInOneCategory = [];
+    return rp(options).then((res) => {
+      const projectsInOneCategory = [];
       const examples = JSON.parse(res);
-      examples.forEach(example => {
+      examples.forEach((example) => {
         let projectName;
         if (example.name === '02_Instance_Container.js') {
-          for (let i = 1; i < 5; i++) {
-            const instanceProjectName = category.name + ': ' + 'Instance Container ' + i;
+          for (let i = 1; i < 5; i += 1) {
+            const instanceProjectName = `${category.name}: Instance Container ${i}`;
             projectsInOneCategory.push({ sketchUrl: example.download_url, projectName: instanceProjectName });
           }
         } else {
           if (example.name.split('_')[1]) {
-            projectName = category.name + ': ' + example.name.split('_').slice(1).join(' ').replace('.js', '');
+            projectName = `${category.name}: ${example.name.split('_').slice(1).join(' ').replace('.js', '')}`;
           } else {
-            projectName = category.name + ': ' + example.name.replace('.js', '');
+            projectName = `${category.name}: ${example.name.replace('.js', '')}`;
           }
           projectsInOneCategory.push({ sketchUrl: example.download_url, projectName });
         }
       });
       return projectsInOneCategory;
-    }).catch(err => {
+    }).catch((err) => {
       throw err;
     });
   }));
 }
 
 function getSketchContent(projectsInAllCategories) {
-  return Q.all(projectsInAllCategories.map(projectsInOneCategory => {
-    return Q.all(projectsInOneCategory.map(project => {
-      const options = {
-        url: project.sketchUrl.replace('?ref=master', '') + '?client_id=' + client_id + '&client_secret=' + client_secret,
-        method: 'GET',
-        headers
-      };
+  return Q.all(projectsInAllCategories.map(projectsInOneCategory => Q.all(projectsInOneCategory.map((project) => {
+    const options = {
+      url: `${project.sketchUrl.replace('?ref=master', '')}?client_id=${clientId}&client_secret=${clientSecret}`,
+      method: 'GET',
+      headers
+    };
 
-      return rp(options).then(res => {
-        const noNumberprojectName = project.projectName.replace(/(\d+)/g, '');
-        if (noNumberprojectName === 'Instance Mode : Instance Container ') {
-          for (let i = 0; i < 4; i++) {
-            let splitedRes = res.split('*/')[1].split('</html>')[i] + '</html>\n';
-            project.sketchContent = splitedRes.replace('p5.js', 'https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.4/p5.min.js');
-          }
-        } else {
-          project.sketchContent = res;
+    return rp(options).then((res) => {
+      const noNumberprojectName = project.projectName.replace(/(\d+)/g, '');
+      if (noNumberprojectName === 'Instance Mode : Instance Container ') {
+        for (let i = 0; i < 4; i += 1) {
+          const splitedRes = `${res.split('*/')[1].split('</html>')[i]}</html>\n`;
+          project.sketchContent = splitedRes.replace('p5.js', 'https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.4/p5.min.js');
         }
-        return project;
-      }).catch(err => {
-        throw err;
-      });
-    }));
-  }));
+      } else {
+        project.sketchContent = res;
+      }
+      return project;
+    }).catch((err) => {
+      throw err;
+    });
+  }))));
 }
 
 function createProjectsInP5user(projectsInAllCategories) {
-  let assetsfiles = [];
   const options = {
-    url: 'https://api.github.com/repos/processing/p5.js-website/contents/dist/assets/examples/assets?client_id=' +
-    client_id + '&client_secret=' + client_secret,
+    url: `https://api.github.com/repos/processing/p5.js-website/contents/dist/assets/examples/assets?client_id=${
+    clientId}&client_secret=${clientSecret}`,
     method: 'GET',
     headers
   };
 
-  rp(options).then(res => {
+  rp(options).then((res) => {
     const assets = JSON.parse(res);
 
     User.findOne({ username: 'p5' }, (err, user) => {
       if (err) throw err;
 
-      async.eachSeries(projectsInAllCategories, (projectsInOneCategory, categoryCallback) => {
-        async.eachSeries(projectsInOneCategory, (project, projectCallback) => {
+      eachSeries(projectsInAllCategories, (projectsInOneCategory, categoryCallback) => {
+        eachSeries(projectsInOneCategory, (project, projectCallback) => {
           let newProject;
           const a = objectID().toHexString();
           const b = objectID().toHexString();
@@ -267,12 +228,12 @@ function createProjectsInP5user(projectsInAllCategories) {
             });
           }
 
-          let assetsInProject = project.sketchContent.match(/assets\/[\w-]+\.[\w]*/g) || project.sketchContent.match(/assets\/[\w-]*/g) || [];
+          const assetsInProject = project.sketchContent.match(/assets\/[\w-]+\.[\w]*/g) || project.sketchContent.match(/assets\/[\w-]*/g) || [];
 
-          assetsInProject.forEach((assetName, i) => {
-            assetName = assetName.split('assets/')[1];
+          assetsInProject.forEach((assetNamePath, i) => {
+            let assetName = assetNamePath.split('assets/')[1];
 
-            assets.forEach(asset => {
+            assets.forEach((asset) => {
               if (asset.name === assetName || asset.name.split('.')[0] === assetName) {
                 assetName = asset.name;
               }
@@ -301,25 +262,61 @@ function createProjectsInP5user(projectsInAllCategories) {
                 children: [],
                 fileType: 'file'
               });
-              console.log('create assets: ' + assetName);
+              console.log(`create assets: ${assetName}`);
               // add asset file inside the newly created assets folder at index 4
               newProject.files[4].children.push(fileID);
             }
           });
 
-          newProject.save((err, newProject) => {
-            if (err) throw err;
-            console.log('Created a new project in p5 user: ' + newProject.name);
+          newProject.save((saveErr, savedProject) => {
+            if (saveErr) throw saveErr;
+            console.log(`Created a new project in p5 user: ${savedProject.name}`);
             projectCallback();
           });
-        }, (err) => {
+        }, (categoryErr) => {
           categoryCallback();
         });
-      }, (err) => {
+      }, (examplesErr) => {
         process.exit();
       });
     });
-  }).catch(err => {
+  }).catch((err) => {
     throw err;
   });
 }
+
+function getp5User() {
+  User.findOne({ username: 'p5' }, (err, user) => {
+    if (err) throw err;
+
+    let p5User = user;
+    if (!p5User) {
+      p5User = new User({
+        username: 'p5',
+        email: 'p5-examples@gmail.com',
+        password: 'test'
+      });
+      p5User.save((saveErr) => {
+        if (saveErr) throw saveErr;
+        console.log(`Created a user p5${p5User}`);
+      });
+    }
+
+    Project.find({ user: p5User._id }, (projectsErr, projects) => {
+      // if there are already some sketches, delete them
+      console.log('Deleting old projects...');
+      projects.forEach((project) => {
+        Project.remove({ _id: project._id }, (removeErr) => {
+          if (removeErr) throw removeErr;
+        });
+      });
+    });
+
+    return getCategories()
+      .then(getSketchesInCategories)
+      .then(getSketchContent)
+      .then(createProjectsInP5user);
+  });
+}
+
+getp5User();
