@@ -1,8 +1,9 @@
 import { Router } from 'express';
-const router = new Router();
 import { renderIndex } from '../views/index';
 import { get404Sketch } from '../views/404Page';
-import { userExists } from '../controllers/user.controller.js';
+import { userExists } from '../controllers/user.controller';
+
+const router = new Router();
 
 // this is intended to be a temporary file
 // until i figure out isomorphic rendering
@@ -48,7 +49,7 @@ router.route('/about').get((req, res) => {
 });
 
 router.route('/:username/sketches').get((req, res) => {
-  userExists(req.params.username, (exists) => (
+  userExists(req.params.username, exists => (
     exists ? res.send(renderIndex()) : get404Sketch(html => res.send(html))
   ));
 });
